+++ packages/myths/src/package.ts
@@ -34,7 +34,7 @@ export const createMythicPackage =
         createSelector:
           <T>(selector: Selector<STATE, T>) =>
             (state: RootState<PKG, STATE>) => {
-              const value = selector(state.__private__[pkg]);
+              const value = selector(state?.__private__?.[pkg]);
               return (value as any).toJS !== undefined
                 ? (value as any).toJS()
                 : value;
